% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_surface_era.R
\name{calculate_surface_area}
\alias{calculate_surface_area}
\title{Calculate Surface Area of a Cone}
\usage{
calculate_surface_area()
}
\value{
The surface area of the cone in square centimeters.
}
\description{
Computes the surface area of an ice cream cone using numerical integration.
}
\examples{
calculate_surface_area()
}
