% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_volume_function.R
\name{calculate_volume}
\alias{calculate_volume}
\title{Calculate Volume of a Cone}
\usage{
calculate_volume()
}
\value{
The volume of the cone in cubic centimeters.
}
\description{
Computes the volume of an ice cream cone using numerical integration.
}
\examples{
calculate_volume()
}
