% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cone Radius Function.R
\name{cone_radius}
\alias{cone_radius}
\title{Cone Radius Function}
\usage{
cone_radius(x)
}
\arguments{
\item{x}{Numeric vector representing the height of the cone.}
}
\value{
A numeric vector with the calculated radii for each x.
}
\description{
Calculates the radius of the cone based on the given piecewise function.
}
\examples{
cone_radius(seq(0, 10, length.out = 100))
}
