% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run Shiny Application}
\usage{
run_app()
}
\value{
None. This function starts the Shiny application.
}
\description{
Launches the Shiny app for weather-based simulations.
}
\examples{
if (interactive()) {
  run_app()
}
}
