% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_party.R
\name{simulate_party}
\alias{simulate_party}
\title{Simulate Party Resources}
\usage{
simulate_party(lambda, prob_one_cone = 0.67, prob_two_cones = 0.33)
}
\arguments{
\item{lambda}{Numeric value for the average number of guests (Poisson distribution mean).}

\item{prob_one_cone}{Probability that a guest eats 1 cone (default: 0.67).}

\item{prob_two_cones}{Probability that a guest eats 2 cones (default: 0.33).}
}
\value{
A list containing the total volume of ice cream and surface area required.
}
\description{
Simulates the total volume and surface area required for a party based on guest behavior.
}
\examples{
simulate_party(lambda = 10, prob_one_cone = 0.67, prob_two_cones = 0.33)
}
